+++ cmake/compiler_setup.cmake
@@ -14,8 +14,8 @@
 
 # C++ Compiler setup
 
-# We use C++11
-set(CMAKE_CXX_STANDARD 11)
+# We use C++14
+set(CMAKE_CXX_STANDARD 14)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 set(CMAKE_CXX_EXTENSIONS OFF)